feat: add event-source-polyfill to support SSE in older browsers
This commit adds the event-source-polyfill library to provide
Server-Sent Events (SSE) support in browsers that do not natively
support it. It also includes the necessary configurations and
modifications to integrate the polyfill into the application.

The changes include:

- Installation of the `event-source-polyfill` and
 `@types/event-source-polyfill` packages.
- Creation of a `polyfills.ts` file to import the polyfill.
- Modification of `main.ts` to conditionally use the native
 EventSource or the polyfill based on availability.
- Update the EventSource constructor in `useXeroAuth.ts` to include
 `withCredentials: true` to allow sending cookies in the SSE
 requests.
- Update tsconfig.json to include polyfills.ts file.

These changes ensure that the application can use SSE functionality
consistently across different browsers, improving the user
experience.