fix(AdminView.vue): fix admin layout to take full screen height
The admin layout was not taking the full screen height, so the
footer was not sticking to the bottom of the page. This commit fixes
this issue by setting the min-height of the admin layout to 100vh and
height to auto. Also, the height of the sidebar and main content was
set to auto to take the full height of the parent.